+++ lint.sh
@@ -1,11 +1,2 @@
-echo "Running bandit..."
-bandit -r app/ skills/ skill_sets/ utils/ -ll
-
-echo "Running pylint..."
-pylint .
-
-echo "Running black..."
-black .
-
-echo "Running isort..."
-isort .
+ruff format
+ruff check --fix